Update README.md
Eight to Seven - Encryption as a Service

Introduction
Welcome to the Eight to Seven GitHub repository! This repository houses the open-source code and resources for the Eight to Seven Encryption as a Service platform. Eight to Seven is a cutting-edge encryption solution that aims to revolutionize data protection and privacy.

About Eight to Seven
At Eight to Seven, we believe that encryption is the key to safeguarding sensitive data in an increasingly digital world. Our platform offers a strong and efficient approach to thwart cyber attacks and protect your valuable information from unauthorized access.

Key Features
High-Speed Quantum Encryption: Our technology enables molecular transformation through DNA, ensuring unparalleled levels of protection.
For the Encrypted data, Our solution employs new generations of encryption algorithms with advanced data archiving encryption enhanced by transformation in the form of high-entropy information bundles.
Smooth Integration: Eight to Seven facilitates secure collaboration and seamless integration with third-party systems across various industries.
Industry-specific Solutions: We provide tailored encryption solutions for industries such as healthcare, retail, automotive, education, fintech, and government.
Contact us for deploying and or support on your upcoming 8)To)7 project

Getting Started
To get started with Eight to Seven, follow these steps:

Demo: Experience the power of Eight to Seven by requesting a demo on the official website.
Full Version: Upgrade to the full version of Eight to Seven for comprehensive data protection and advanced encryption capabilities.
Contributing
We welcome contributions from the open-source community to enhance the Eight to Seven platform. If you have any bug reports, suggestions, or feature requests, please open an issue or submit a pull request.

License
This project is licensed under the GPL License. Feel free to explore and examen the code for your own projects or contribute to the Eight to Seven platform.
We will also open two communication channels
1-Public
2-Semi public (ask to join if you are a Researcher or Professor in Cryptography who wants to contribute to our code.)


This a a new way of Encryption so for all your questions 
Contact
For more information about Eight to Seven, visit our website or contact us directly at quanty22@eighttoseven.com.
We consider 8(to)7 as an authentic 100% post-quantum encryption